sulong: Type size is now long, introduce Type#*Exact methods and TypeOverflowException
(cherry picked from commit c85923fdaa1fbd5e8f94c0b3dbd49ef009e149a1)

+++ sulong/projects/com.oracle.truffle.llvm.runtime/src/com/oracle/truffle/llvm/runtime/nodes/memory/LLVMGetStackSpaceInstruction.java
@@ -46,13 +46,13 @@ import com.oracle.truffle.llvm.runtime.types.Type;
 
 public abstract class LLVMGetStackSpaceInstruction extends LLVMExpressionNode {
 
-    protected final int size;
+    protected final long size;
     protected final int alignment;
     protected final Type symbolType;
 
     @CompilationFinal private FrameSlot stackPointer;
 
-    public LLVMGetStackSpaceInstruction(int size, int alignment, Type symbolType) {
+    public LLVMGetStackSpaceInstruction(long size, int alignment, Type symbolType) {
         this.size = size;
         this.alignment = alignment;
         this.symbolType = symbolType;
@@ -73,18 +73,18 @@ public abstract class LLVMGetStackSpaceInstruction extends LLVMExpressionNode {
 
     public abstract static class LLVMGetStackForConstInstruction extends LLVMGetStackSpaceInstruction {
 
-        public LLVMGetStackForConstInstruction(int size, int alignment, Type symbolType) {
+        public LLVMGetStackForConstInstruction(long size, int alignment, Type symbolType) {
             super(size, alignment, symbolType);
         }
 
         @CompilationFinal(dimensions = 1) private Type[] types = null;
-        @CompilationFinal(dimensions = 1) private int[] offsets = null;
+        @CompilationFinal(dimensions = 1) private long[] offsets = null;
 
         public void setTypes(Type[] types) {
             this.types = types;
         }
 
-        public void setOffsets(int[] offsets) {
+        public void setOffsets(long[] offsets) {
             this.offsets = offsets;
         }
 
@@ -92,7 +92,7 @@ public abstract class LLVMGetStackSpaceInstruction extends LLVMExpressionNode {
             return types;
         }
 
-        public int[] getOffsets() {
+        public long[] getOffsets() {
             return offsets;
         }
 
@@ -104,7 +104,7 @@ public abstract class LLVMGetStackSpaceInstruction extends LLVMExpressionNode {
 
     public abstract static class LLVMAllocaConstInstruction extends LLVMGetStackForConstInstruction {
 
-        public LLVMAllocaConstInstruction(int size, int alignment, Type symbolType) {
+        public LLVMAllocaConstInstruction(long size, int alignment, Type symbolType) {
             super(size, alignment, symbolType);
         }
 
@@ -124,7 +124,7 @@ public abstract class LLVMGetStackSpaceInstruction extends LLVMExpressionNode {
 
         private final UniqueSlot uniqueSlot;
 
-        public LLVMGetUniqueStackSpaceInstruction(int size, int alignment, Type symbolType, UniqueSlot uniqueSlot) {
+        public LLVMGetUniqueStackSpaceInstruction(long size, int alignment, Type symbolType, UniqueSlot uniqueSlot) {
             super(size, alignment, symbolType);
             this.uniqueSlot = uniqueSlot;
         }
@@ -143,7 +143,7 @@ public abstract class LLVMGetStackSpaceInstruction extends LLVMExpressionNode {
     @NodeChild(type = LLVMExpressionNode.class)
     public abstract static class LLVMAllocaInstruction extends LLVMGetStackSpaceInstruction {
 
-        public LLVMAllocaInstruction(int size, int alignment, Type symbolType) {
+        public LLVMAllocaInstruction(long size, int alignment, Type symbolType) {
             super(size, alignment, symbolType);
         }
 
@@ -151,7 +151,7 @@ public abstract class LLVMGetStackSpaceInstruction extends LLVMExpressionNode {
         protected LLVMNativePointer doOp(VirtualFrame frame, int nr,
                         @CachedLanguage LLVMLanguage language) {
             try {
-                return LLVMNativePointer.create(LLVMStack.allocateStackMemory(frame, language.getLLVMMemory(), getStackPointerSlot(), size * (long) nr, alignment));
+                return LLVMNativePointer.create(LLVMStack.allocateStackMemory(frame, language.getLLVMMemory(), getStackPointerSlot(), size * nr, alignment));
             } catch (StackOverflowError soe) {
                 CompilerDirectives.transferToInterpreter();
                 throw new LLVMAllocationFailureException(this, soe);
